+++ powersync/src/error.rs
@@ -3,7 +3,7 @@ use std::io;
 use std::sync::Arc;
 use std::{borrow::Cow, fmt::Display};
 
-use http_client::http_types::StatusCode;
+use http::StatusCode;
 use rusqlite::Error as SqliteError;
 use rusqlite::types::FromSqlError;
 use thiserror::Error;
@@ -35,9 +35,16 @@ impl From<serde_json::Error> for PowerSyncError {
     }
 }
 
-impl From<http_client::Error> for PowerSyncError {
-    fn from(value: http_client::Error) -> Self {
+impl From<http::Error> for PowerSyncError {
+    fn from(value: http::Error) -> Self {
         RawPowerSyncError::Http { inner: value }.into()
+    }
+}
+
+#[cfg(feature = "reqwest")]
+impl From<reqwest::Error> for PowerSyncError {
+    fn from(value: reqwest::Error) -> Self {
+        RawPowerSyncError::Reqwest { inner: value }.into()
     }
 }
 
@@ -82,17 +89,25 @@ pub(crate) enum RawPowerSyncError {
     /// Wraps `serde_json` errors.
     #[error("Internal error while converting JSON: {inner}")]
     JsonConversion { inner: serde_json::Error },
+    #[error("Error decoding stream from sync service: {desc}")]
+    SyncServiceResponseParsing { desc: &'static str },
     /// Used when a backend connector returns an invalid URI as a service URL.
     #[error("Invalid PowerSync endpoint: {inner}")]
     InvalidPowerSyncEndpoint { inner: url::ParseError },
     /// HTTP errors while downloading sync lines.
     #[error("HTTP error: {inner}")]
-    Http { inner: http_client::Error },
+    Http { inner: http::Error },
     /// A generic IO error that occurred in the SDK.
     #[error("IO error: {inner}")]
     IO {
         #[from]
         inner: io::Error,
+    },
+    #[cfg(feature = "reqwest")]
+    #[error("HTTP error: {inner}")]
+    Reqwest {
+        #[from]
+        inner: reqwest::Error,
     },
     #[error("The PowerSync service did not accept credentials returned by connector")]
     InvalidCredentials,